Update validateInvocation recording. setupRumbleRecording uses defaults
This change removes the uses of intercept in validateInvocation tests.
Using intercept did not create a good test. This commit also resolves
an issue with testing where intercept and/or recording setup was causing
errors during testing ci. There are two new tests. One for a
'happy path' and one for an error during the validateInvocation.